requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
einops==0.8.1
filelock==3.13.1
fsspec==2024.6.1
idna==3.4
imagecodecs==2024.9.22
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==25.0
pillow==11.0.0
pip==25.1
pytorch-msssim==1.0.0
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.28.1
scikit-image==0.25.2
scipy==1.15.2
setuptools==78.1.1
sympy==1.13.3
tifffile==2024.12.12
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.13
wheel==0.45.1
zipp==3.22.0
